DeleteDeployment is using InstanceDeleter

+++ bosh-director/lib/bosh/director/api/snapshot_manager.rb
@@ -51,9 +51,10 @@ module Bosh::Director
         result
       end
 
-      def self.delete_snapshots(snapshots)
+      def self.delete_snapshots(snapshots, options={})
+        keep_snapshots_in_the_cloud = options.fetch(:keep_snapshots_in_the_cloud, false)
         snapshots.each do |snapshot|
-          Config.cloud.delete_snapshot(snapshot.snapshot_cid)
+          Config.cloud.delete_snapshot(snapshot.snapshot_cid) unless keep_snapshots_in_the_cloud
           snapshot.delete
         end
       end
